Add dependencies for froala and font-awesome
Add a bower dependency for froala-wysiwyg-editor and an npm dependency
for the ember-cli-font-awesome addon. That way this addon only has to
import the froala editor, but relies on the font-awesome addon for that
requirement. Besides, many projects already use the font-awesome addon
anyway. Add these dependancies to both the project and addon (for
development and testing).