Major rewrite of event handling
Breaking changes!!! The `content` attribute is no longer updated by the
editor, users should use an event action to `(mut)` the property
instead. Changing `options` no longer cause the editor to
re-initialize. Finally, the `event` argument has been removed from
event action callbacks since it is not available when adding event
handlers directly to the editor. All of this needs to be detailed in
release notes and possibly the README.

Closes #4 #16, #17, and #18